+++ Client-Alias/Client-Alias/Main/RoomInfo/RoomInfoRouter.swift
@@ -3,14 +3,17 @@ import UIKit
 protocol RoomInfoRouterInput {
 
     func showUserActions(user: String,
+                         isAdmin: Bool,
                          addHandler: @escaping () -> Void,
-                         adminHandler: @escaping () -> Void,
-                         deleteHandler: @escaping () -> Void)
-    func showTeams()
-    func showMembers()
+                         adminHandler: @escaping () -> Void?,
+                         deleteHandler: @escaping () -> Void?)
+    func showTeams(participantID: UUID, teams: [TeamInfo])
+    func showMembers(model: MembersModel)
     func showAddTeam(roomId: UUID)
     func showGameSettings()
     func showTeamActions(team: String, deleteHandler: @escaping () -> Void)
+    func showAlert(title: String, message: String)
+    func showAlert()
 //    func showInviteCode(code: String)
 }
 
@@ -25,24 +28,27 @@ final class RoomInfoRouter {
 extension RoomInfoRouter: RoomInfoRouterInput {
 
     func showUserActions(user: String,
+                         isAdmin: Bool,
                          addHandler: @escaping () -> Void,
-                         adminHandler: @escaping () -> Void,
-                         deleteHandler: @escaping () -> Void) {
+                         adminHandler: @escaping () -> Void?,
+                         deleteHandler: @escaping () -> Void?) {
         let vc = UIAlertController(title: user, message: "", preferredStyle: .actionSheet)
-        let addAction = UIAlertAction(title: "Изменить команды", style: .default) { _ in
+        let addAction = UIAlertAction(title: "Добавить в команду", style: .default) { _ in
             addHandler()
         }
         vc.addAction(addAction)
         
-        let adminAction = UIAlertAction(title: "Сделать администратором", style: .default) { _ in
-            adminHandler()
+        if !isAdmin {
+            let adminAction = UIAlertAction(title: "Сделать администратором", style: .default) { _ in
+                adminHandler()
+            }
+            vc.addAction(adminAction)
+            
+            let deleteAction = UIAlertAction(title: "Удалить", style: .destructive) { _ in
+                deleteHandler()
+            }
+            vc.addAction(deleteAction)
         }
-        vc.addAction(adminAction)
-        
-        let deleteAction = UIAlertAction(title: "Удалить", style: .destructive) { _ in
-            deleteHandler()
-        }
-        vc.addAction(deleteAction)
         let cancelAction = UIAlertAction(title: "Отменить", style: .cancel)
         vc.addAction(cancelAction)
         view?.present(vc, animated: true)
@@ -60,14 +66,14 @@ extension RoomInfoRouter: RoomInfoRouterInput {
         view?.present(vc, animated: true)
     }
 
-    func showTeams() {
-        let vc = TeamsModuleConfigurator().configure().view
+    func showTeams(participantID: UUID, teams: [TeamInfo]) {
+        let vc = TeamsModuleConfigurator().configure(participantID: participantID, teams: teams).view
         vc.sheetPresentationController?.detents = [.medium()]
         view?.present(vc, animated: true)
     }
 
-    func showMembers() {
-        let vc = MembersModuleConfigurator().configure().view
+    func showMembers(model: MembersModel) {
+        let vc = MembersModuleConfigurator().configure(model: model).view
         view?.navigationController?.pushViewController(vc, animated: true)
     }
 
@@ -80,6 +86,15 @@ extension RoomInfoRouter: RoomInfoRouterInput {
     func showGameSettings() {
         let vc = GameSettingsModuleConfigurator().configure().view
         view?.navigationController?.pushViewController(vc, animated: true)
+    }
+    
+    func showAlert(title: String, message: String) {
+        let alert = UIAlertController.makeProblemAlert(title: title, message: message, anchoredBarButtonItem: .none)
+        view?.present(alert, animated: true, completion: nil)
+    }
+    
+    func showAlert() {
+        view?.present(UIAlertController.makeProblemAlert(anchoredBarButtonItem: .none), animated: true, completion: nil)
     }
     
 //    func showInviteCode(code: String) {
